Move botanist.GenerateDNSProviderName to gutil.GenerateDNSProviderName
