+++ pkg/apis/core/validation/shoot.go
@@ -27,8 +27,8 @@ import (
 	"github.com/gardener/gardener/pkg/apis/core/helper"
 	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
 	"github.com/gardener/gardener/pkg/features"
-	"github.com/gardener/gardener/pkg/operation/botanist"
 	"github.com/gardener/gardener/pkg/utils"
+	gutil "github.com/gardener/gardener/pkg/utils/gardener"
 	cidrvalidation "github.com/gardener/gardener/pkg/utils/validation/cidr"
 	versionutils "github.com/gardener/gardener/pkg/utils/version"
 
@@ -451,7 +451,7 @@ func validateDNS(dns *core.DNS, fldPath *field.Path) field.ErrorList {
 		idxPath := fldPath.Child("providers").Index(i)
 
 		if provider.SecretName != nil && provider.Type != nil {
-			providerName := botanist.GenerateDNSProviderName(*provider.SecretName, *provider.Type)
+			providerName := gutil.GenerateDNSProviderName(*provider.SecretName, *provider.Type)
 			if names.Has(providerName) {
 				allErrs = append(allErrs, field.Invalid(idxPath, providerName, "combination of .secretName and .type must be unique across dns providers"))
 				continue
